make sure you pass arch

+++ install.sh
@@ -1,5 +1,11 @@
-#This is for internal testing only
+if [ -z "$1" ]
+    then 
+    echo 'You must provide ARCH as first parameter'
+    echo './install.sh x86_64-linux-gnu'
+    exit;
+fi
 
+ARCH=$1
 #Remove any old version
 docker-compose down 
 sudo rm -rf volumes
@@ -13,5 +19,5 @@ mkdir volumes/rtl_datadir
 mkdir volumes/tor_datadir
 mkdir volumes/tor_servicesdir
 mkdir volumes/tor_torrcdir
-docker-compose build --build-arg ARCH=x86_64-linux-gnu
+docker-compose build --build-arg ARCH=($ARCH)
 docker-compose up -d
